# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API de agendamento do VS Instituto de Beleza
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Staff login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /admin/settings:
    put:
      operationId: updateSettings
      tags: [admin]
      summary: Update salon information
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalonSettings"
      responses:
        "200":
          description: Saved settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonSettings"
  /admin/services:
    get:
      operationId: listAdminServices
      tags: [admin]
      summary: All services including inactive
      security: [{ session: [] }]
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminService"
    post:
      operationId: createService
      tags: [admin]
      summary: Create a service
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminService"
  /admin/services/{id}:
    patch:
      operationId: updateService
      tags: [admin]
      summary: Update a service
      security: [{ session: [] }]
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminService"
  /admin/professionals:
    get:
      operationId: listAdminProfessionals
      tags: [admin]
      summary: All professionals including inactive
      security: [{ session: [] }]
      responses:
        "200":
          description: Professionals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProfessional"
    post:
      operationId: createProfessional
      tags: [admin]
      summary: Create a professional
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessionalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfessional"
  /admin/professionals/{id}:
    patch:
      operationId: updateProfessional
      tags: [admin]
      summary: Update a professional
      security: [{ session: [] }]
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessionalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfessional"
  /admin/accounts:
    get:
      operationId: listAccounts
      tags: [admin]
      summary: Staff login accounts
      security: [{ session: [] }]
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffAccount"
    post:
      operationId: createAccount
      tags: [admin]
      summary: Create a staff login
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffAccountCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
  /admin/accounts/{id}:
    patch:
      operationId: updateAccount
      tags: [admin]
      summary: Update a staff login
      security: [{ session: [] }]
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffAccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: vs_session
  parameters:
...
    AppointmentId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Professional:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        initials: { type: string }
        startsAt: { type: string }
        endsAt: { type: string }
        weekdays:
          type: array
          items: { type: integer }
      required: [id, name, initials, startsAt, endsAt, weekdays]
...
    StaffUser:
      type: object
      properties:
        role:
          type: string
          enum: [professional, admin]
        professional:
          oneOf:
            - $ref: "#/components/schemas/Professional"
            - type: "null"
        displayName: { type: string }
        email: { type: string }
      required: [role, professional, displayName, email]
...
    LoginInput:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    PasswordChangeInput:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
    SalonSettings:
      type: object
      properties:
        name: { type: string }
        whatsapp: { type: string }
        address: { type: string }
        instagram: { type: string }
        about: { type: string }
        salonSharePercent: { type: number }
      required: [name, whatsapp, address, instagram, about, salonSharePercent]
    AdminService:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        durationMinutes: { type: integer }
        active: { type: boolean }
      required: [id, name, price, durationMinutes, active]
    ServiceInput:
      type: object
      properties:
        name: { type: string }
        price: { type: number }
        durationMinutes: { type: integer }
        active: { type: boolean }
    AdminProfessional:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        initials: { type: string }
        startsAt: { type: string }
        endsAt: { type: string }
        weekdays:
          type: array
          items: { type: integer }
        active: { type: boolean }
      required: [id, name, initials, startsAt, endsAt, weekdays, active]
    ProfessionalInput:
      type: object
      properties:
        name: { type: string }
        initials: { type: string }
        startsAt: { type: string }
        endsAt: { type: string }
        weekdays:
          type: array
          items: { type: integer }
        active: { type: boolean }
    StaffAccount:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [professional, admin]
        professionalId: { type: ["integer", "null"] }
        active: { type: boolean }
      required: [id, name, email, role, professionalId, active]
    StaffAccountCreate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role:
          type: string
          enum: [professional, admin]
        professionalId: { type: ["integer", "null"] }
      required: [name, email, password, role]
    StaffAccountUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role:
          type: string
          enum: [professional, admin]
        professionalId: { type: ["integer", "null"] }
        active: { type: boolean }